+++ src/Traits/IsCabinetKitUser.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Posio\CabinetKit\Traits;
+
+trait IsCabinetKitUser
+{
+    use HasAccount;
+    use HasCustomFields;
+    use HasSettings;
+}