+++ tailwind-preset.cjs
@@ -1,0 +1,3 @@
+module.exports = {
+    content: ['./vendor/posio/cabinet-kit/resources/js/**/*.vue'],
+};